Allow a setting for a separate USB console
The Orin AGX board provides access to the kernel and UEFI consoles
via USB only today, not via the 40-pin header.

To allow jumpstarter accessing systems like these, we can enable
the jumpstarter software to look-up for a separate usb console
on the host.

i.e. with an Orin AGX board we could store:
set-config usb_console TOPOD83B461B-if01

this would in turn, tell the jumpstarter software on the host
to look up for devices in /dev/serial/by-id/ that contain
such string, like:

/dev/serial/by-id/usb-NVIDIA_Tegra_On-Platform_Operator_TOPOD83B461B-if01

and use that when a Console() is requested to the DUT